Avoid crashes when deformation is called with non-final parameters
These are functions that have ambition to store their results in internal
tables for rapid later retrieval. But the tables are designed to store data
only for final parameters, which would crash after the computation at a
non-final parameter if there was any computed data worth storing. The
essential deformation computation is recursive and the recursive calls already
automatically involve final parameters, so that is made a precondition, and
each recursive call does its own look-up and storage. To not expose the user
to the technical limitation to final, the wrapper functions now expand their
argument into a list of final ones (using |extended_finalise| in the case of
twisted deformation) and sum over the results obtained for those.