Make internal |is_normal| test |z==normal(z)|; reorder parameter printing
The |is_normal| is simplified to do the right thing. It is only used for
printing a qualification for parameters, where we gave the "zero" and "not
(semi)final" conditions priority to it. So finally "non-normal" will only
print for nonzero semifinal parameters, and the former change will not matter.