Change |branch| command so that it will accept non-final standard parameters
There is no reason it should refuse non-finals, since the code was able to
handle this already. In fact it could probably even handle non-standard
parameters, since it is calling |KhatContext::standardize| internally which
would rewite a non-standard |StandardRepK| using Hecht-Schmid identities.

A version of |branch| operating directly on a |ParamPol| was added, and in
basic.at versions for getting a single K_type. Polished K_type.at slightly.