New functions, changes to existing ones, other small changes, in extParamPol.at
in deform_unreduced, if p=(x,lambda,nu) is not delta-fixed, the old
behavior was to replace the extended parameter (x,lambda,nu;0) with
(x,lambda,0;0). This is OK for the application to Hermitian forms; it
ignores some terms (1+s)*P, which will be multiplied by (1-s), giving
0. However it is better in the big world to keep track of these extra
terms, which it now does (by eliminating the if type=0 conditional)

Change int:sign to int:type in deform_unreduced for consistency

in change_nu, if (x,lambda,nu;0) is an extended parameter of induced
type, it could be that (x,lambda,nu*r) is now delta-fixed (this always
happens if r=0). Now there is a test to replace such a parameter
with a pair (x,lambda,nu*r,\pm1)

removed a few print debugging lines (commented and otherwise)

character_formula now has a third int:type argument

added c_form_std_big and c_form_irreducible_as_sum_of_standards

replaced (ParamPol,ParamPol,ParamPol) with ExtParamPol in a number of places