Rewrite |deformation_terms| to use vector rather than polynomial internally
Since this function is operating on a single block (in fact on the set of
final block elements in the interval up to the entry element) there is no need
to use polynomials other than those stored in |KL_list|; the linear algebra
for the inversion operation is easily done in coordinates on those block
elements. For speed a sparse map |index| from hash codes of |StandardRepr|
values back to coordinate positions is used in converting the polynomials.
The result, which is greatly simplified, appears to be also noticeably faster.