Remove the |witness| parameter to |SRK_context::isFinal| and friends
Having a references passed to store additional output information may be the
right thing to do in principle, it is annoying to have to to provide a
variable to accept that output for each call (especially from the debugger).
So in this change we provide a class-static variable that will hold the
additional result and a method to fetch and clear it.