Correct several issues with |finals_below|, and make |expand_finals| use it
An important issue was a call |result.erase(rit++)| where |rit| was an
itererator into the list |result|; although |erase| removes the node |rit|
"points to" (i.e., after the point where |insert| would insert), the iterator
remains valid and gets to automatically point to the next node; this is unlike
iterators for say |std::list| where the argument to |erase| get invalidated by
the operation. This is a caveat that should accompany |containers::sl_list|.

Other changes were replacing two instances of |z| by |*rit|, the element of
current interest rather than the initial parameter, and constucting a new
|StandardRepr| on complex descents rather than just modifying the |x_part| and
|y_bits| (even though maybe the |height| of the parameter does not actually
change, that was certainly not a consideration when this code was written).