Remove post-increment operators from list iterators, lacking any useful purpose
Using post-increment for iterators into linked STL containers is useful,
notably to pass the result as argument to |erase| while retaining a valid
iterator. For the linked containers implemented in sl_list.h however, it is
the iterator following the argument to |erase| that gets invalidated, so using
post-increment here is certainly not useful (the passed iterator remains
valid, and now accesses the node after the erased one, if any). Since there is
no reason to use it with |insert| either (the increment taking place before
the insertion, the iterator would have skipped a node after the inserted one;
this is no different than for STL iterators), it seems better to suppress the
operator altogether, thus removing a trap for users. On the other hand weak
iterators, which cannot be used with erase/insert, retain their post-increment.