Fix |Cayley@(vec,Param)| to leave result integrally dominant; improve repr.cpp
The method |W_act| used to reconstruct a parameter with the same |gamma| as
the initial one was implemented wrongly, but it was doomed anyway since
keeping a non integrally dominant |gamma| could imply transforming a standard
parameter into a non standard parameter, which is not what Cayley should do.

In the end the |W_act| function was simple not called, and indeed removed.

Some independent reorganisation of repr.cpp for transparency was also applied