Changed finalize@Param to return a ParamPol
finalize(p) used to return a [Param]. It is better (and simpler)
to return a ParamPol, and in fact this is simply the same thing
as ParamPol:p.

I changed all of the places in the at scripts where finalize was
called to accomodate this change (there were only a few).