Make tests for twist-fixed parameters be applied only after normalizing them
A parameter should be considered fixed under $\delta$ if twisting by $\delta$
results in a parameter _equivalent_ to the original one. We used to directly
test |p==twist(p)|, but this can give false negatives, either when the
parameter is non-dominant, or else if twisting its involution $\theta$ results
in a different but conjugate involution. It seems that both defects can be
resolved by performing |make_dominant| and then moving to the canonical
involution for the singular subsystem; twisting a parameter there results in
another dominant one at a canonical involution, and this can only be
equivalent to that parameter when it is actually equal to it. This is
implemented in |is_twist_fixed| and called by the relevant wrapper functions.