Revise |add| method to overload table, separating out a checking function
The idea is to separate out detection of a potential conflict from actual
modification of the table. Thus by doing the testing first, combined
definitions can be rejected up front as a whole when there is a problem,
rather than breaking off their execution half way through when a conflicting
case is encountered while already adding items to the tables.