Cleanup work in utilities subdirectory. Use |Poset::Elt| rather than |set::Elt|
In fact the module 'set', a lone file set.h only defining the types |set::Elt|
and |set::EltList|, is entirely removed. In practice such sets are at least
endowed with a poset structure, and the requirement for being able to store
that practically limit use to sets of less than 2^32 elements, for the
numbering of which |size_t| is unnecessarily wide. Therefore we now instead
define |Poset::Elt| as |unsigned int|, saving space notably in Hasse diagrams.