Creating new branch to work on cyclotomic fields
Fields and operations are working. The main issue is
inverse. This is implemented using the Galois theory
idea. However it has fairly serious integer overflow
problems. I was not careful about this, for example
in flex_add etc. around line 26 in rational_matrices.at.
There is some easy cleaning up to do, but I'm not sure
this is sufficient to make this work for reasonable size n.