Weyl character formula, preliminary version
First working version of Weyl character formula, evaluated at element
of finite order. The result is an element of a cyclotomic field; these
fields are implemented in cycltomic.at.

rational_polynomial.at: polynomials over Q
cyclotomic.at: cycltomic fields
weyl_character_formula.at Weyl character formula
finite_dimensional_signature.at signature of finite dimensional representations,
an application of the Weyl character formula
coxeter.at  the coxeter element

An application of the WCF is that trace(pi(x)), where x is a strong
real form, gives the signature of the invariant Hermitian form on a
finite dimensional representation. In this case we've implmented a
special version, wcf_strong_real, which takes advantage of features of
this case, and returns a number of the form k*exp(2i\pi/n) for k an
integer. For this application we do the calculation using rational
numbers (which are up to 2^63-1), rather than integers (which are
limited to 2^31-1).

The calculation of the Weyl character formula is limited by integer
overflow problems. Currently the |int| type is at most 2^31-1.
We may implement arbitrary precision integers to get around this.