Correct serious flaw: multiply-and-subtract is harder than multiply-and-add
The loop updating the remainder to a smaller one was written like the inner
loop of the multiplication function, but this can produce borrows outside the
|two_digits| value used which were ignored. As a remedy we reorganised the
loop so that it differs from the multiplication loop only by complimenting
of values coming from and returning to the remainder, so in particular the
product is being _added_, and can be made to never overflow the |two_digits|
accumulator. This does require addition and subtraction cases to be separated.