Change parsing process for integer denotations allowing for very large values
Integer constants used to be converted to an integer value in the lexical
analyser, but now it stores a string which will only be converted to a value
when the executable denotation is built, after type-checking. Thus the result
can be a |big_int| and does not suffer from truncation to 32 (or 64) bits.