Fix |int_format|, and estimate of space in |big_int| constructed from string
In |int_format|, the shared pointer itsel rather than the |big_int| held was
being preinted.

In contructing |big_int| from a string of characters, there was an error in
estimating the required number of |digit| words, but this was without
consequences since only affected the size passed to |reserve|. For decimal
numbers, it divided the string length by 4 rather than by 32/4=8 to estimate
the number of words necessary, reserving for large numbers twice the space.