Further changes to Weyl character formula, and small changes to cyclotomic.at
The formula was ignoring the central torus.  In fixing this, I decided
to write some new functions for creating new root data, together with
change of basis |ratmat|. These are in group_operations.at:

  adjoint_info (RootDatum rd) = (RootDatum,mat):
  simply_connected_cover_info(RootDatum rd)=(RootDatum, ratmat):
  simply_connected_cover(RootDatum rd)=RootDatum:
  simply_connected_cover_derived_info(RootDatum rd)=(RootDatum,ratmat):
  simply_connected_cover_derived(RootDatum rd)=RootDatum:

Also I needed left_inverse of a |ratmat|, which is now in ratmat.at.

Added functions to cycltomic.at:

  complex_conjugate(cyclotomicFieldElement z)=cyclotomicFieldElement:
  set bar(cyclotomicFieldElement z)=cyclotomicFieldElement:complex_conjugate(z)
  set is_real(cyclotomicFieldElement z)=bool:z=z.bar
  set absolute_value_squared(cyclotomicFieldElement z)=cyclotomicFieldElement:z*z.bar

Small changes necessary so that all scripts load, and fix whitespace errors:

  group_operations.at changed ss_rank to semisimple_rank
  ratmat.at added <matrix.at