Fix bug affecting integer division: (right) shifts should be unsigned
These shifts are applied only to remainders for a division, and before maybe
changing it to negative, so on unsigned shift is in place, and was counted on.

By the way divisors are sclaed, this bug would preduce a wrong sign remainder
from long division quite frequently (when scaled remainder had "sign" bit set).