Simplify big integer division by avoiding global shift of dividend and divisor
For conceptual simplicity, left bit-shifts were applied to dividend and
divisor to make the leading bit of the latter a 1, and consequently the
remainder had to be right-shifted again. But it suffices to compute the
leading two words of the left-shifted divisor, and to compute a corresponding
leading part of the left-shifted current remainder each time a quotient digit
|q_hat| needs to be computed (this is exercise 4.3.1:37 of Knuth's TAOCP).

In passing the computation of |q_hat| is siplified by using only the two
(rather than three) leading words of the left-shifted current remainder when
deciding whether to decrement the initial |q_hat| guess; this makes slightly
more conservative the decision taken, but the case where no decrementation is
decides still easily excludes the scenario where for the correctness of the
algorithm a decrementation is necessary (namely when |q_hat| overshoots by 2).