Tweek integer division: avoid extending dividend when one can avoid it easily
We just need to make sure the initial "remainder" (the dividend) satsifies the
general invariant for remainders, which is that in absolute value it lies
between minus the divisor (inclusive) and plus the divisor (exclusive). When
inspection of the leading words says this is clearly true, which happens about
half the time on average, we can avoid finding an initial quotient digit 0.
We take care however not to suppress the only quotient digit to be produced.