Strengthen the ratvec code, and also make some cosmetic improvements in passing
One constructor forgot to make the denominator positive, and flipping the sign
of the denominator when a negative denominator is passed can use |negate|. We
now use |big_int| in rational vector comparisons to avoid any overflow during
such operations. The matrix reduction code still forgot to convert integers to
the template type |C| in several places, which is now fixed.