Use branch labels more interestingly, to possibly reorder branches as needed
Users can now specify the collection of branches in any order, which will then
be untangled into the proper order, but the collection still needs to be
complete. In particular if any variants of a defined union type are missing
names, there is no way to make a valid discrimination expression for the type.

A next step is to allow a default for missing branches.