Fix bug in big integer division: ensure the quotient comes out with proper sign
This was already affirmed in the comments, but was not ensured by the initial
code that just set up the invariant necessary for the quotient digits to be
found. A twice as strict intial condition (ensured by sign extending if
necessary) will ensure that the sign bit of the quotient will be correct.