Change occurrences of |last_param| into |first_param|, as ordering was reversed
This should really have been done at commit 34894a9f (March 20, 2017) when the
ordering of polynomials was made to be by increasing height, which is more or
less the reverse of what it was before. In some cases it is important to get
hold of the "leading monomial" of a ParamPol, in |composition_series| for
instance, and this now is done using |first_param| rather than |last_param|.
In other cases it is just important to get any monomial present in the
polynomial, and here typically |first_param| was already used; these instances
have been left unchanged. Only one (recent) instance of |last_param| remains.