Make |solve| return (vec|void) rather than [vec], and adapt client scripts
The linear algebra solving functions are a good example of how unions can be
used: they used to return the type [vec] just to return either 1 or 0 vectors
(where the former _did not_ indicate uniqueness of the solution), since one
should be able to communicate "no solution" to the caller without an error.
The union type (vec|void) is more suited to this purpose, with the second
variant indicating that no solution exists. we chose the name |maybe_a_vec|
for the union type, with tags (injector functions) |solution|, |no_solution|.
As the latter takes 0 arguments, its name make have to be made more specific
later, if other union types are added for "no solution" to other questions.

Handling the results of the call has to be adapted to this change. Besides the
general case .. | .. | .. esac expression, we provide functions |any| to just
test existence of a solution (for |assert|), and |requisition| (with error
message) to assert then necessity to find a solution in certain instances.

To accomodate situations where existence of a solution