Fix bug in bigint subtraction (|sub_from|): forgot to transfer |x| to |*this|
In the case where |sub| needs to do nothing because either subtracting a digit
|0| without carry or a digit |~0| with carry (neither the current digit nor
the carry changes), the corresponding code of |sub_from| still has to transfer
the unchaged digit form the argument |x| to |*this|, which was forgotten.

Since axis subtraction ctually uses |sub_from| rather than |sub|, this led to
an embarassing change of large integer numbers when subtracting 0 from them.