Fix error with |root_index| which returned its negative values as huge instead
This error caused in partuclar real induction code to malfunction, but of
course potentially all clients of |root_index|. The technical reason for the
error is that the index was obtained by subtracting from the internal root
index the number of positive roots (so that positive roots start at 0), and
both these quantities being unsiged, the result was an unsigned quatity (in
fact unsigned long) that for negative roots was very near the upper limit of
values for that unsigned type, namely 2^64. This caused no problem as long as
the value was converted to a signed machine integer type anyway, but with the
advent of big integers, and in particular with its constructors taking into
account the true signed or unsigned interpretation of the passed argument, the
inxed values got an offset 2^64 for negative roots. The solution made was to
cast both quantities subtracted to a signed integer type first before
perfoming the subtraction, and then converting the result to |big_int|.

In passing the result type for the |numPosRoots| method and friends was
chenged from |unsigned long| to |rootNbr| (which is |unsigned short), in
accordance with the intended meaning (and assuming the number of roots will
never exceed 2^16=65536 does not seem in any way a dangerous assumtion).