Make some small refinements: conversion |[[rat]->ratmat|, |highest_root|
The former is done in ratmat.at (which file is to be loaded only by those
accepting personal risk for overflow), the latter in basic.at. Related to the
latter matter, the documentation of |dual@RootDatum| now warns that for it,
the last root might not be the highest one.