Add coercion vec->[rat] so |vec| overloads come before |[rat]|; adapt basic.at
Only one recently added definition in basic.at gave a conflict, and (after its
removal) another related defintiion in extended.at ceased to give an error.
The error occurred because a |[rat]| overload was tried where an |[int]|
argument (non list display) was provided, which could not be converted; the
intended |vec| overload was present but too far down, and was not tried.