Fix glitch in |is_close|: void type should be tested before primitive type
It was not intended that void would compare close to any other type, but due
to the formulation for primitice types, the tested close to void because of
the possible voiding coercion. This now being fixed, calls of functions
expectig void with a primitive type argument no longer succeed as they used to
(with a voiding to dispose of the primitive type argument). Note that calls
without overloading (though a local or global function type variable) still
do allow voiding in the argument position if the function takes no arguments.