Prepare for making types capable of internal type naming scheme (for recursion)
A variant |type_name| is added to the union in |expr_type|, and a static
member is added that stores association of the integer value staored for that
variant with a type identifier and a type expression. There is no possiblity
yet of actually creating values of this variant, but methods do cater for it.