Change argument convention for |list| in basic.at, and order in coded bitsets
Following the general principle that argument ordering with the simpler things
first is preferred, and for coherence with |all|, |first| and friends, the
limit number for |list|, and its companion |complement|, is made first
argument, and the predicate second. Also, since bitsets are coded for passing
to |list| or |complement|, they now put the number before the condition.

Also define a |complement| for explicit strictly increasing lists of integers