Thoroughly overhaul the script lattice.at, changing leaving few things intact
Nearly all implementations have been revised to use |echelon|, not |diagonalize|
or |adapted_basis|. Many have been renamed, a few whose function was not clear
have been removed, and most of those that remain were simplified. One central
new function is |adapted_direct_sum|, which computes a basis for the saturated
image of a linear map, and a of a complementary direct factor. Most other
functions like linear maps induced in a quotient space are defined in terms of
this decomposition and the bases for the factors it chooses. It should be
noted that many functions are more about subspaces, represented by saturated
lattices, than about general $\Z$-modules.