Complete (hopefully) the work on |defined_type_mapping::add_typedefs|
This is surprisingly long, and completely untested for now. Also the way in
which all of this is integrated into |type_expr| will probably need
modification, both to separate cases where we want |kind()==type_name| to be
visible and transparent, and to protect the public base object (with type
bindings) of |type_expr::type_map|, which would seem appropriate.