Repair more errors with tabled types: expand if needed, and correct print test
The access methods like |tuple()| needed to handle table types by expanding,
but did not. The printng of types should use a name only if one was provided,
but the test that was in place just tested the type was tabled at all.

In passing |compon_variant| was renamed |row_variant|, which is shorter.