Fix some bugs in typedef code: |merge_sort| argument and typedef expansion
The list argument to |merge_sort| should of course be by reference, as
modifying it is the whole point; that was forgotten. Also, because the typedef
code will not handle one named type being defind as another named type, the
client code needs to expand the top level definition in this case for safety.