Simplify error reporting by defining template method |error_base::operator<<|
Thus we can create an erro object, then write information to its message as if
writing to an object of some class like |std::ostringstream| that is derived
from |std::ostream|; such objects are no longer needed.

Exploiting this involves a lot of changes; for now only global.w is done.