Replace |typedef_table| by values stored inside |type_expr::type_map|
This change makes the |type_table| class obsolete, and it is removed. We also
removed |type_expr::remove| and the unique call to, because removing the name
of a recursive type could lead to non-terminating printing (moreover the removed
call was not in the right place; it belonged outside the conditional it was in).