Revise error types so that |operator<<| retains type of the error expression
This was a misunderstanding of the |throw| mechanism, where the type thrown is
the static type of the error expression, not the dynamic (derived) type of the
error object. Therefore if the error expression is an application of a method
defined in a base class, the type thrown cannot be the derived type of the error
object from which this method is called (this applied to |operator<<|). The
remedy applied is to define the (non virtual!) method at the level of the
derived class, and return a reference to the derived class from it.