Tune type output: don't rename primitives, change projector/injector printing
We disable storing a name for primitive types when tabled: there could be at
most one alternative name for a primitive, and there is no reasonable mechanism
to decide whether the primitive or alternative name would be more appropriate.

In storing types for projector and injector functions, we use the tabled number
for the tuple or union type, which will print as the type name if one was given.

Finally we have changed output for projector and injector values themselves.