Remove default constructor for |PreRootDatum|, and the reason it was needed.
This is part of the more general project of removing default constructors for
objects that don't need to ever be in undefined state. Removal will require
replacing functions without output parameters of the mentioned type to functions
returning a value of the type, as was done with a |getInteractive| instance (now
renamed to |get_pre_root_datum|). In the testrun unit, an unused default
constructor for |CoveringIterator| also had to be removed for this purpose.