Replace (co)weight lists in |PreRootData|, and its methods, by |LatticeMatrix|
It is more compact to store a matrix than a list of weights eo coweights, and
exporting them from a |PreRootDatum| either individually or as a matrix makes
most client code simpler (though in some cases it had to be rewritten a bit).
The only place we still use these lists for |PreRootData| is in arguments to its
constructor; after that everything now has been made matrix based. This is also
in preparation for hashing pre-rootdata to enforce unique storage of root data.