Add coroot preference atribute to all |RootSystem| constructing functions
Versions that default this attribute were defined in basic.at, so that scripts
continue to function correctly. Also the attribute is propagated in the implicit
converion from |RootDatum| to |PreRootDatum|, and a test function included.
Inside the library the defualt argument is removed; explicit values are given.
Finally atlas-function.help was updated to document the new specifications.