Start stamping out |clone|: changed |get_own| and |force_own|
The purpose is to not force atlas user types that don't need it to implement
|clone|, and allow them to not have a copy constructor at all. The current
solution is to make those that do want to copy have a public copy constructor
(having a private copy constructor as was used by |clone| is very hard to
reconcile with using |std::make_shared|, as we wish to), which |get_own| calls.

The main remaining client of |clone| is |uniquify|. It could in principle also
invoke copy constructors explicitly since the actual type is known when it is
used. This require modification of |component_assignment::assign|; for later.