Make |uniquify| templated, and have it return a downcast raw pointer
Rather than depend on a virtual |clone| method, the uses of |uniquify| can all
tell it the actual type held in the |shared_value|, so that duplication (if
needed) can use |std::make_shared| together with a copy constructor. We can take
advantage of returning a downcast raw pointer to non-const, which is what the
caller will need anyway when invoking |uniquify|; some call sites are therefore
simplified. We are now ready to remove |clone| and unneeded copy constructors.