Finally remove those |ref_count| and |dual| fields from |inner_class_value|
These were vestiges of a home-baked reference counting scheme (introduced long
ago, before we were even aware of |std::shared_ptr|) whose purpose was to avoid
creating identical |InnerClass| objects by operations in the atlas programming
language, with an additional twist to ensure that taking the dual twice (which
amounts to creating a copy anyway) would be detectable and result in equality
when tested in an atlas program. Now a new scheme is in place that ensures that
no two |inner_class_value| objects can coexist with the exact same mathematical
properties (among which we count preference for roots or coroots in the root
datum), but rather identical constructions will lead to one shared object.

These fields were ugly (because an |InnerClass| can already give almost all
information its dual can, and reference counted sharing is now ubiquitous in the
atlas program due to using |std::shared_ptr|), but somewhat hard to stamp out,
because the presence of the |dual| field was being (ab)used in several places.
In the end we needed to add a test for |inner_class_value|s being duals.

After this commit several cleaning up operations can and should be done; notably
the test for equality of inner classes can now be done by pointer comparison at
the |inner_class_value| level (as the temporary function |identity|, which is
intended only for testing the current changes, currently does).