Make |inner_class@(LieType,[ratvec],string)| into script function in basic.at
This built-in function was awkward, and was the sole manner of getting an inner
class without using the |inner_class_value::build| method. The script version is
quite simple, and should present no observable differences (and anyway, few seem
interested in this 'Fokko' method of building inner classes anymore). Also
equality testing of inner classes is now done by testing identity of the
|inner_class_value| objects; it should present no observable differences either.