Added functionality to compute conjugacy class of given WeylElt
This works in many cases, but is probably *not* correct in general.
I am only checking the Levi factors agree, and the characteristic
polynomials agree. Probably you have to check the characteristic
polynomials agree simple-term by simple-term. (Also this is ignoring
the one ambiguity in F4).

Fixed a bug in simple_factors.at: at each step need to merge all
terms with non-trivial inner product with the given one.