Extend syntax for choice clauses, and allow all of them after |while|
We now allow more flexible branch order in |if| and integer |case| clauses
(though giving |else| before |then| is only admitted in the former case). We
also extended the possibilities of a 'do_expr' (the entity between |while| and
the corresponding |od|) to allow all choice expressions (union-case and
discrimination expression used to be disallowed), allow parentheses around them,
and allow the branch order relaxation of this commit within a 'do_expr' as well.