Change |rootdata::wrt_distinguished| to reverse order and twist result
Though at glance insinificant, this was a very hard thing to set straight. There
was an incoherence between comment and code with respoect to the direction in
which the result word was being generated; in the end the order is indeed
opposite to the one in |factor_dominant|, but in addition a twist by the
distinguished involution (discovered at the end of the function) was also
needed, which explains why in practice, when the result is a twisted involution,
forgetting to do both had no effect. I (hopefully) cleaned up all other places
in the code and comments for which this function is relevant, removing from
compilation one place that was in fact unused (in subsytem.cpp).