Fix bug in |ratvec| comparison causing erratic behaviour: forgot to test length
For |ratvec| values of different length, comparison of entries was done without
testing that their numbers were equal, causing existing entries to be compared
with nonexisting ones (out of bounds) if the second argument was shorter than
the first, causing uncertain results. This was causing the while loop in the now
removed function |partition_of_w_BCD| to, in very rare cases that the spurious
tests returned "equal", terminate before having reduced its polynomial to [1]/1.