Fixed ratvec versus vec issue for complex orbits
An SL(2) triple has a semisimple element which is a ratvec.  This is
the image, under a map phi:sl(2)->Lie(G), of
[1/2,-1/2]=alpha^v/2=rho^vee for SL(2) (for these coordinates use
GL(2)).

The functions referring to H in complex_nilpotent_orbits are referring
to 2*this element. This is phi(alpha^v), and is in X_*, and is
therefore represented by a vec in atlas.

The function H_candidates@RootDatum returns [ratvec], for example for
SL(3). However the subset which pass the test to be in all_H@RootDatum
are all vecs.

Conclusion: S_weight and H_candidates return ratvecs, but all_H
returns [vec]. Also the type ComplexNilpotent is (RootDatum, vec).

Note: if (rd,v) is a ComplexNilpotent, then <v,alpha>\in {0,1,2} for
all simple roots alpha. (This fact in itself only implies that v is in
the coweight lattice, which is not contained in X^*.)