Rename |poly| to |i_poly|, and adapt (and polish) the scripts that use it
The type |poly_mat| was also changed to |i_poly_mat|; no other renaming done.

The name change is towards the goal of having various types of (univariate)
polynomials the user can choose from; for instance there will be |int_poly| as
well, giving polynomials with coefficients not limited to machine integers.
The name |i_poly| (only in types) is also easier in searching than just |poly|

This was an occasion to do some long due cleaning up for readability. Some code
was rewritten in the process, so there is a slight risk of breaking something