Change built-in interface to LieType to facilitate getting simple factors
Now |simple_factors@Lie_type| is built-in, and contrary to the old |%@Lie_type|
it does not report torus factors. The possibility of torus factors in a list of
otherwise simple type codes was annoying and error prone. The old functionality
is now available as |factors@Lie_type| in basic.at (but hardly used). Some more
changes to the LieType section of basic.at were made, and all scripts that were
using |%| were adapted to call either |factors| or |simple_factors| instead.