Rework the wreps_type_C.at script, reinterpreting the Noel-Jackson paper
The main new point is working with a single set of partitions with one part
singled out, rather than two separate sets, one of just partition and another
with a part labelled (and included), as it used to be. More comments on what
exactly is oging on are still needed, as well as cleaning up older stuff.