Fix |is_valid@(string,vec)|: do not refuse partitions due to invalid Lie types
For instance |is_valid("B",[3])| should return |true|, but it threw an error by
trying to pass |Lie_type("B1")| to |is_valid@(LieType,vec)|, but that Lie type
refuses to be built. Maybe |Lie_type@string| should be more lenient by producing
type A1 when asked for B1, but that would be of no help here, since the set of
partitions (of 3) valid for B1 differs from those (partitions of 2) for A1.