Rewrote |realredgp::minimal_torus_part| to fix a synthetic real form bug
The problem was that |real_form(InnerClass:G,involution(x),torus_factor(x))|
would return a real form different from |G| in general (and in particular for
|G=SU(3,1)| and |x=KGB(G,3)|). It was caused by |realredgp::minimal_torus_part|
returning a torus part not in the same orbit as the one implicitly provided (by
the difference between the square class cocharacter and the provided torus
factor). This in turn was caused by the assumption that the orbit (for the
imaginary weyl group at the fundamental fiber) could be read off from the fiber
partition for the relevant square class, stored in the fundamental fiber.
However, that partition was computed (during construction of the fundamental
|Fiber| object) using as base point for the affine imaginary Weyl group action
the adjoint fiber element coset representative for the square class, while the
actual orbit uses the elected (by |some_coch|) cocharacter for the square class,
computed from the actual "square value" characterisiing the square class. The
two do not define the same base grading, so the orbits do not match; moreover
the problem is hard to avoid, since no coset representative for the square class
is available when |some_coch| does its work.

The solution was to not use stored orbits (or partitions) at all, but to do the
orbit generation "by hand" in |realredgp::minimal_torus_part|. As a consequence
the |InnerClass::torus_parts_for_grading_shift| method which erroneously tried
to find torus part candidates is no longer used, and in fact it was removed,
because by what was just said it cannot be made to work correctly without adding
more arguments (it would need access to the cocharacter grading involved).

So there is still discrepancy between grading from |some_coch| for a square
class |csc| and |G.fundamental_fiber().grading(class_base(scs))| as used in
|Fiber::makeStrongReal|; it would probably be best to do without any of the
strong real form classification in |Fiber| objects altogether.