Classify set_type commands in scripts into simple or general type definitions
Here simple means without the brackets, which gives just a type name expanded on
input, while general means potentially recursive, used in type equivalencing,
and the name will be used on output whenever the equivalence was discovered.

This is a bit subjective, as most type definitions could be simple ones; however
for elaborate tuples with field selectors, using the abbreviation on output
seems to be desirable, so we used general definitions for those types.