Remove some legacy occurrences of |rat| by |int|; bits are no issue any more
It used to be the case that |rat| had 64 bits for the numerator (and the
denominator as well) where |int| had only 32, and for this reason some integer
values were represented as rationals (with denominator 1). Since we now have
unlimited size |int| values, this is no longer pertinent, and we reverted to the
more natural |int| type in those places. There was some subsequent
simplification of code, and also finite_dimensional_signature.at was cleaned up.